Clear Range Of Bits
Method: clearInRange
The clearInRange method is used to clear a range of bits within a given integer n, starting from the index i to the index j (both inclusive). It returns the integer with the specified range of bits cleared while preserving the rest of the bits.

Approach:
Create Bitmasks:
Create two bitmasks a and b to represent the bits to be cleared and preserved, respectively.
Bitmask a: Set all bits from index j+1 to the most significant bit to 1, leaving the lower bits as 0.
Bitmask b: Set all bits from index 0 to i to 1, leaving the higher bits as 0.
Combine Bitmasks:
Combine the two bitmasks using a bitwise OR operation to create the final bitmask c.
This final bitmask will have 1 in the range [i, j] and 0 elsewhere.
Clear Bits:
Perform a bitwise AND operation between the original number n and the final bitmask c to clear the bits in the specified range.
Return Result:
Return the result obtained after clearing the specified range of bits.

Example:
Input: clearInRange(10, 2, 4)
n: 10 (1010 in binary)
i: 2, j: 4
Bitmasks:
a: 11100000
b: 00000011
c: 11100011 (OR of a and b)

Result:
n & c: 1000 (binary) => 8 (decimal)
This method offers a straightforward way to clear a range of bits within an integer, making it useful for various bitwise operations and manipulations.