Count_set_Bits
Description:
The Count_set_Bits class provides a method countSetBits to calculate the number of set bits (bits with value 1) in an integer. It iteratively checks each bit of the integer using bitwise operations.

Approach:
The countSetBits method iterates over each bit of the given integer using a while loop. It checks if the least significant bit (LSB) is set by performing a bitwise AND operation with 1. If the result is not zero, it means that the LSB is set, and the method increments the count of set bits. Then, it right shifts the integer by one position to examine the next bit. This process continues until all bits are examined.

Example:
Input: countSetBits(16)
n: 16 (10000 in binary)
Operation:
Iteration 1: 10000 & 00001 = 00000 (LSB not set)
Iteration 2: 01000 & 00001 = 00000 (LSB not set)
Iteration 3: 00100 & 00001 = 00000 (LSB not set)
Iteration 4: 00010 & 00001 = 00000 (LSB not set)
Iteration 5: 00001 & 00001 = 00001 (LSB set)
Result: Number of set bits = 1
This approach provides an efficient way to count the number of set bits in an integer using bitwise operations.