Fast Exponentiation
The Fast Exponentiation Algorithm efficiently calculates the power of a number using bitwise operations. It reduces the number of multiplications required by exploiting the binary representation of the exponent.


Method:
Input Parameters: The algorithm takes two integers, a (the base) and n (the exponent).
Initialization: Initialize a variable ans to store the result, initially set to 1.
Bitwise Iteration: Iterate through the bits of the exponent n using a while loop.
Bitwise AND Operation: Check if the least significant bit of n is set (i.e., n & 1 == 1).
Multiplication: If the current bit is 1, multiply the result ans by the base a.
Base Squaring: Square the base a for the next iteration.
Right Shift: Shift the exponent n to the right (equivalent to dividing by 2).
Return Result: After iterating through all the bits of the exponent, return the final result ans.

Approach:
By iterating through the binary representation of the exponent, the algorithm effectively reduces the number of multiplications required.

It exploits the fact that raising a number to a power involves repeated multiplication, and it takes advantage of the binary representation of the exponent to perform fewer multiplications.
The algorithm achieves a time complexity of O(log n) due to its iterative approach, making it highly efficient for calculating large exponent values.